+++ docs/source/changes.rst
@@ -17,6 +17,7 @@ v0.3
   overriding :classmethod:`~formfields.do_display_data`.
 * Dropped support for Python 3.2. Nobody is really using it and it's a pain to
   integrate other libraries.
+* Support for `django-simple-captcha <https://github.com/mbi/django-simple-captcha>`_
 
 
 v0.2